feat: implement keybindings listener and disabler composables in use-keybindings hook
